+++ lib/src/json/sounds/custom_sound.g.dart
@@ -11,13 +11,14 @@ CustomSound _$CustomSoundFromJson(Map<String, dynamic> json) => CustomSound(
           $enumDecode(_$CustomSoundAssetStoreEnumMap, json['assetStore']),
       id: json['id'] as String,
       gain: (json['gain'] as num?)?.toDouble() ?? 0.7,
-    );
+    )..audioBusId = json['audioBusId'] as String?;
 
 Map<String, dynamic> _$CustomSoundToJson(CustomSound instance) =>
     <String, dynamic>{
       'id': instance.id,
       'gain': instance.gain,
       'assetStore': _$CustomSoundAssetStoreEnumMap[instance.assetStore],
+      'audioBusId': instance.audioBusId,
     };
 
 const _$CustomSoundAssetStoreEnumMap = {